Revert "Begin swapping to dataclasses for config verifications"
