+++ requirements.txt
@@ -2,4 +2,3 @@ Jinja2==3.0.2
 PyYAML==5.3.1
 PyPDF2==1.26.0
 roman==3.3
-
